rename boardgame folder to game
