+++ src/game/entities/score.ts
