+++ src/game/phases/setup.ts
