+++ src/ui/GameState.ts
@@ -1,6 +1,6 @@
 import { State } from "boardgame.io";
-import { WizardState } from "../boardgame/WizardState";
-import { PlayerID } from "../boardgame/entities/players";
+import { WizardState } from "../game/WizardState";
+import { PlayerID } from "../game/entities/players";
 
 export interface GameState extends State {
   wizardState: WizardState;
