+++ src/ui/WizardBoard.tsx
@@ -5,8 +5,8 @@ import { GameContext } from "./GameContext";
 import { HeaderBar } from "./header/HeaderBar";
 import { OpponentsContainer } from "./player/OpponentsContainer";
 import { Table } from "./table/Table";
-import { PlayerID } from "../boardgame/entities/players";
-import { WizardState } from "../boardgame/WizardState";
+import { PlayerID } from "../game/entities/players";
+import { WizardState } from "../game/WizardState";
 import { ScorePad } from "./score/ScorePad";
 import { FinalScoreModal } from "./gameover/FinalScoreModal";
 import { GameState } from "./GameState";
